requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Django==5.1
djangorestframework==3.15.2
drf-yasg==1.21.7
fonttools==4.53.1
greenlet==3.0.3
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pluggy==1.5.0
polars==1.2.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyarrow==17.0.0
pyparsing==3.1.2
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.31
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
